requirements.txt
pymorphy2==0.8
python-telegram-bot==12.2.0
PyYAML==5.2
SQLAlchemy==1.3.12
matplotlib==3.2.0
palettable==3.3.0
Pillow==7.0.0